4-3.db/2_graph_neo4j/ifc_to_neo4j_parser/requirements.txt
ifcopenshell
neo4j
python-dotenv